removed println and replaced with ui.printMessage

+++ src/SharedElements/Document.java
@@ -107,7 +107,7 @@ public class Document {
     
     public void setDate(String date) {this.dateCreated = date;}
     
-    public void setPrice(int price) {this.price = price;}
+    public void setPrice(double price) {this.price = price;}
     
     public void setQuantity(int quantity) {this.availableAmount = quantity;}
